feat: Add basic CLI support + file renaming
